refactor: Refactor and rename classes and methods for clarity
The class and method names related to the parsing of sign descriptions have been modified for better readability and clarity. Class `RpaSignDescriptionParser` is renamed to `RpaSignDescParser` to simplify the name. Other ancillary methods and classes are also renamed, and redundant methods are removed to enhance the readability and encapsulation of the code. Additionally, NullPointerException is prevented by adding a null-check for input string parameters.